+++ NEWS.md
@@ -1,4 +1,7 @@
-## lineup2 0.3-2 (2021-01-16)
+## lineup2 0.3-3 (2021-01-18)
+
+- Added function `plot_sample()` for plotting a selected row or column
+  from a distance/similarity matrix.
 
 - Fixed typo in help file for `dist_betw_arrays()`.